Housing: Fix wire formats and budget values from captured packet data
- Fix NotifyPermissionsFailure Result field from uint32 to uint16
- Fix budget values to match game data: interior 910-2050 by level,
  exterior 200, room 19, fixture 1000 (old formulas were 12-67x low)
- Populate WorldServerInfo housing GUIDs (House, OwnerAccount,
  CosmeticOwner, Neighborhood) at login
- Fix NeighborhoodPlayerEnterPlot to single PlayerGuid (remove PlotGuid)
- Fix NeighborhoodPlayerLeavePlot to empty body (0 bytes)
- Redesign NeighborhoodRosterResidentUpdate as count+array of
  PlayerGuid+PlotIndex, update all 5 handler call sites
- Add neighborhood_map and neighborhood_plot data for Map 2736

+++ src/server/game/Handlers/HousingHandler.cpp
@@ -1107,7 +1107,7 @@ void WorldSession::HandleHousingSvcsTeleportToPlot(WorldPackets::Housing::Housin
     if (!neighborhood)
     {
         WorldPackets::Housing::HousingSvcsNotifyPermissionsFailure response;
-        response.Result = static_cast<uint32>(HOUSING_RESULT_NEIGHBORHOOD_NOT_FOUND);
+        response.Result = static_cast<uint16>(HOUSING_RESULT_NEIGHBORHOOD_NOT_FOUND);
         SendPacket(response.Write());
         return;
     }
@@ -1117,7 +1117,7 @@ void WorldSession::HandleHousingSvcsTeleportToPlot(WorldPackets::Housing::Housin
     if (!mapData)
     {
         WorldPackets::Housing::HousingSvcsNotifyPermissionsFailure response;
-        response.Result = static_cast<uint32>(HOUSING_RESULT_INVALID_PLOT);
+        response.Result = static_cast<uint16>(HOUSING_RESULT_INVALID_PLOT);
         SendPacket(response.Write());
         return;
     }
@@ -1185,7 +1185,7 @@ void WorldSession::HandleHousingSvcsStartTutorial(WorldPackets::Housing::Housing
 
         // Notify client of failure
         WorldPackets::Housing::HousingSvcsNotifyPermissionsFailure failResponse;
-        failResponse.Result = static_cast<uint32>(HOUSING_RESULT_NEIGHBORHOOD_NOT_FOUND);
+        failResponse.Result = static_cast<uint16>(HOUSING_RESULT_NEIGHBORHOOD_NOT_FOUND);
         SendPacket(failResponse.Write());
         return;
     }
